move `ExecutionReport` to own file, sort table more nicely
