+++ core/src/runtime/mod.rs
@@ -5,6 +5,7 @@ mod opcode;
 mod program;
 mod record;
 mod register;
+mod report;
 mod state;
 mod syscall;
 #[macro_use]
@@ -16,18 +17,16 @@ pub use opcode::*;
 pub use program::*;
 pub use record::*;
 pub use register::*;
+pub use report::*;
 pub use state::*;
 pub use syscall::*;
 pub use utils::*;
 
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
-use std::fmt::{Display, Formatter, Result as FmtResult};
 use std::fs::File;
-use std::hash::Hash;
 use std::io::BufWriter;
 use std::io::Write;
-use std::ops::{Add, AddAssign};
 use std::sync::Arc;
 
 use thiserror::Error;
@@ -91,86 +90,7 @@ pub struct Runtime {
     pub report: ExecutionReport,
 
     /// Whether we should write to the report.
-    pub should_report: bool,
-}
-
-/// Integer type used to count instructions and syscalls for `ExecutionReport`.
-type Count = u64;
-
-#[derive(Default, Debug, Clone, PartialEq, Eq)]
-pub struct ExecutionReport {
-    pub instruction_counts: HashMap<Opcode, Count>,
-    pub syscall_counts: HashMap<SyscallCode, Count>,
-}
-
-impl ExecutionReport {
-    pub fn total_instruction_count(&self) -> Count {
-        self.instruction_counts.values().sum()
-    }
-
-    pub fn total_syscall_count(&self) -> Count {
-        self.syscall_counts.values().sum()
-    }
-}
-
-/// Helper function. "Derives" `+=` if the value type of a `HashMap` supports it.
-fn hashmap_add_assign<K, V>(lhs: &mut HashMap<K, V>, rhs: HashMap<K, V>)
-where
-    K: Eq + Hash,
-    V: AddAssign,
-{
-    for (k, v) in rhs.into_iter() {
-        // Can't use `and_modify(...).or_insert(...)` because we want to use `v` in both places.
-        match lhs.entry(k) {
-            Entry::Occupied(e) => *e.into_mut() += v,
-            Entry::Vacant(e) => drop(e.insert(v)),
-        }
-    }
-}
-
-// Used to concatenate two `ExecutionReport`s.
-impl AddAssign for ExecutionReport {
-    fn add_assign(&mut self, rhs: Self) {
-        hashmap_add_assign(&mut self.instruction_counts, rhs.instruction_counts);
-        hashmap_add_assign(&mut self.syscall_counts, rhs.syscall_counts);
-    }
-}
-
-// Used to concatenate two `ExecutionReport`s.
-// `AddAssign` will probably be used more. `Add` is here for completeness.
-impl Add for ExecutionReport {
-    type Output = Self;
-
-    fn add(mut self, rhs: Self) -> Self::Output {
-        self += rhs;
-        self
-    }
-}
-
-impl Display for ExecutionReport {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        writeln!(f, "Instruction Counts:")?;
-        let mut sorted_instructions = self.instruction_counts.iter().collect::<Vec<_>>();
-
-        // Sort instructions by opcode name
-        sorted_instructions.sort_by_key(|&(opcode, _)| opcode.to_string());
-        for (opcode, count) in sorted_instructions {
-            writeln!(f, "  {}: {}", opcode, count)?;
-        }
-        writeln!(f, "Total Instructions: {}", self.total_instruction_count())?;
-
-        writeln!(f, "Syscall Counts:")?;
-        let mut sorted_syscalls = self.syscall_counts.iter().collect::<Vec<_>>();
-
-        // Sort syscalls by syscall name
-        sorted_syscalls.sort_by_key(|&(syscall, _)| format!("{:?}", syscall));
-        for (syscall, count) in sorted_syscalls {
-            writeln!(f, "  {}: {}", syscall, count)?;
-        }
-        writeln!(f, "Total Syscall Count: {}", self.total_syscall_count())?;
-
-        Ok(())
-    }
+    pub print_report: bool,
 }
 
 #[derive(Error, Debug)]
@@ -231,7 +151,7 @@ impl Runtime {
             emit_events: true,
             max_syscall_cycles,
             report: Default::default(),
-            should_report: false,
+            print_report: false,
         }
     }
 
@@ -652,9 +572,9 @@ impl Runtime {
         let lookup_id = create_alu_lookup_id();
         let syscall_lookup_id = create_alu_lookup_id();
 
-        if self.should_report && !self.unconstrained {
+        if self.print_report && !self.unconstrained {
             self.report
-                .instruction_counts
+                .opcode_counts
                 .entry(instruction.opcode)
                 .and_modify(|c| *c += 1)
                 .or_insert(1);
@@ -874,7 +794,7 @@ impl Runtime {
                 b = self.rr(Register::X10, MemoryAccessPosition::B);
                 let syscall = SyscallCode::from_u32(syscall_id);
 
-                if self.should_report && !self.unconstrained {
+                if self.print_report && !self.unconstrained {
                     self.report
                         .syscall_counts
                         .entry(syscall)
@@ -1054,7 +974,7 @@ impl Runtime {
     /// Execute up to `self.shard_batch_size` cycles, returning the events emitted and whether the program ended.
     pub fn execute_record(&mut self) -> Result<(ExecutionRecord, bool), ExecutionError> {
         self.emit_events = true;
-        self.should_report = true;
+        self.print_report = true;
         let done = self.execute()?;
         Ok((std::mem::take(&mut self.record), done))
     }
@@ -1062,7 +982,7 @@ impl Runtime {
     /// Execute up to `self.shard_batch_size` cycles, returning a copy of the prestate and whether the program ended.
     pub fn execute_state(&mut self) -> Result<(ExecutionState, bool), ExecutionError> {
         self.emit_events = false;
-        self.should_report = false;
+        self.print_report = false;
         let state = self.state.clone();
         let done = self.execute()?;
         Ok((state, done))
@@ -1094,14 +1014,14 @@ impl Runtime {
 
     pub fn run_untraced(&mut self) -> Result<(), ExecutionError> {
         self.emit_events = false;
-        self.should_report = true;
+        self.print_report = true;
         while !self.execute()? {}
         Ok(())
     }
 
     pub fn run(&mut self) -> Result<(), ExecutionError> {
         self.emit_events = true;
-        self.should_report = true;
+        self.print_report = true;
         while !self.execute()? {}
         Ok(())
     }
@@ -1259,7 +1179,7 @@ pub mod tests {
             use super::Opcode::*;
             use super::SyscallCode::*;
             super::ExecutionReport {
-                instruction_counts: [
+                opcode_counts: [
                     (LB, 10723),
                     (DIVU, 6),
                     (LW, 237094),
